pyproject.toml
[tool.poetry]
name = "musicbot"
version = "0.1.0"
description = "Matrix music bot"
authors = ["Benjamin Herman <benjamin@metanomial.com>"]
readme = "readme.md"

[tool.poetry.dependencies]
aiohttp = "^3.8.3"
python = "^3.10"
python-dotenv = "^0.21.1"
schedule = "^1.1.0"
matrix-nio = "^0.20.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
pre-commit = "^3.0.4"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
